@@ -105,63 +111,41 @@
       "last 1 chrome version"
     ]
   },
-  "dependencies": {
-    "@date-io/moment": "^1.3.13",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/pickers": "^3.3.10",
-    "axios": "^0.21.1",
-    "clsx": "^1.1.1",
-    "electron-log": "^4.3.2",
-    "file-type": "^16.3.0",
-    "lodash": "^4.17.21",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-intl": "^5.15.5",
-    "react-markdown": "^5.0.3",
-    "react-quill": "^1.3.5",
-    "react-redux": "^7.2.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "sharp": "^0.27.2",
-    "uuid": "^8.3.2"
-  },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.10",
-    "@testing-library/react": "^10.4.9",
-    "@testing-library/user-event": "^10.4.1",
-    "@types/electron-devtools-installer": "^2.2.0",
-    "@types/file-type": "^10.9.1",
-    "@types/jest": "^25.2.3",
-    "@types/lodash": "^4.14.168",
-    "@types/node": "^14.14.37",
-    "@types/react-dom": "^17.0.3",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/sharp": "^0.27.3",
-    "@types/uuid": "^8.3.0",
-    "@typescript-eslint/eslint-plugin": "^4.20.0",
-    "@typescript-eslint/parser": "^4.20.0",
-    "concurrently": "^5.3.0",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^14.0.0",
+    "@types/electron-devtools-installer": "^2.2.1",
+    "@types/jest": "^29.0.0",
+    "@types/node": "^16.11.19",
+    "@types/react-redux": "^7.1.22",
+    "@typescript-eslint/eslint-plugin": "^5.9.1",
+    "@typescript-eslint/parser": "^5.9.1",
+    "concurrently": "^7.0.0",
     "copyfiles": "^2.4.1",
     "cross-env": "^7.0.3",
-    "electron": "^12.0.2",
-    "electron-builder": "^22.10.5",
-    "electron-devtools-installer": "^3.1.1",
-    "eslint": "^7.23.0",
-    "eslint-plugin-promise": "^4.3.1",
-    "eslint-plugin-react": "^7.23.1",
-    "husky": "^4.3.8",
-    "node-sass": "^4.14.1",
-    "typescript": "^4.2.3",
-    "wait-on": "^5.3.0"
+    "electron": "^21.1.0",
+    "electron-builder": "^23.0.3",
+    "electron-devtools-installer": "^3.2.0",
+    "eslint": "^8.6.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "husky": "^8.0.1",
+    "sass": "^1.54.8",
+    "typescript": "^4.5.4",
+    "wait-on": "^6.0.0"
   },
-  "husky": {
-    "hooks": {
-      "pre-push": "npm run lint"
-    }
+  "dependencies": {
+    "@reduxjs/toolkit": "^1.8.1",
+    "axios": "^1.1.0",
+    "clsx": "^1.1.1",
+    "electron-log": "^4.4.4",
+    "i18next": "^21.6.6",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-i18next": "^11.15.3",
+    "react-redux": "^8.0.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0"
   }
 }
