@@ -1,5 +1,5 @@
 import { BrowserWindow } from 'electron';
 
 export default (browserWindow: BrowserWindow): void => {
-  browserWindow.webContents.send('appSetPlatform', process.platform);
+  browserWindow.webContents.send('setPlatform', process.platform);
 };
